==> experiments/storm_events/01_download_docs.ipynb <==
import os
import requests

# Download the Storm Events list of files
BRONZE_PATH = '/usr/datalake/bronze/stormevents/docs'
BRONZE_FEDCODES_PATH = '/usr/datalake/bronze/fedcodes'
NCEI_URL = 'https://www.ncei.noaa.gov/pub/data/swdi/stormevents/csvfiles/'
FEDCODES_URL = 'https://prd-tnm.s3.amazonaws.com/StagedProducts/GeographicNames/FederalCodes/FedCodes_AllStates_Text.zip'

os.makedirs(BRONZE_PATH, exist_ok=True)
os.makedirs(BRONZE_FEDCODES_PATH, exist_ok=True)

filename = f'{BRONZE_PATH}/csvfiles.html'

if os.path.exists(filename):
    print(f"File already exists: {filename}")
else:
    response = requests.get(NCEI_URL)
    response.raise_for_status()

    with open(filename, 'wb') as f:
        f.write(response.content)

    print(f"File downloaded successfully to {filename}")

filename = f'{BRONZE_FEDCODES_PATH}/FedCodes_AllStates_Text.zip'

if os.path.exists(filename):
    print(f"File already exists: {filename}")
else:
    response = requests.get(FEDCODES_URL)
    response.raise_for_status()

    with open(filename, 'wb') as f:
        f.write(response.content)

    print(f"File downloaded successfully to {filename}")
