==> experiments/storm_events/02_transform_csvfiles.ipynb <==
import os
import re
import pandas as pd

# Transform the csvfiles index HTML into a CSV
BRONZE_PATH = '/usr/datalake/bronze/stormevents/docs'
SILVER_PATH = '/usr/datalake/silver/stormevents/docs'
REGEX_FILENAME = "StormEvents_details-ftp_v1\\.0_d\\d{4}_c\\d{8}\\.csv\\.gz(?=\\\")"

os.makedirs(SILVER_PATH, exist_ok=True)

filename = f'{BRONZE_PATH}/csvfiles.html'
with open(filename, 'r', encoding='utf-8') as file:
    html = file.read()

filename = f'{SILVER_PATH}/csvfiles.csv'
df = pd.DataFrame(re.findall(REGEX_FILENAME, html), columns=['filename'])
df['year'] = df['filename'].str[30:34].astype(int)
df['updated'] = pd.to_datetime(df['filename'].str[36:44])
df.to_csv(filename, index=False)
df.head()
