==> experiments/storm_events/02_transform_fedcodes.ipynb <==
import glob
import zipfile
import pandas as pd

# Unzip the fedcodes, combine files, and create a csv
BRONZE_FEDCODES_PATH = '/usr/datalake/bronze/fedcodes'
SILVER_FEDCODES_PATH = '/usr/datalake/silver/fedcodes'

zipfile_path = f'{BRONZE_FEDCODES_PATH}/FedCodes_AllStates_Text.zip'

with zipfile.ZipFile(zipfile_path, 'r') as zip_ref:
    zip_ref.extractall(SILVER_FEDCODES_PATH)

glob_path = f'{SILVER_FEDCODES_PATH}/Text/FederalCodes_*.txt'

# Get CSV files list from a folder
csv_files = glob.glob(glob_path)

# Read each CSV file into DataFrame
# This creates a list of dataframes
df_list = (pd.read_csv(file, delimiter='|') for file in csv_files)

# Pandas doesn't like to concatinate "empty" DataFrames
df_list = [dataframe for dataframe in df_list if dataframe.shape[0] > 0]

# Concatenate all DataFrames
df = pd.concat(df_list, ignore_index=True)

del df_list
df.head()

csv_path = f'{SILVER_FEDCODES_PATH}/FederalCodes.csv'

df.to_csv(csv_path, index=False)
