==> experiments/storm_events/03_download_files.ipynb <==
import os
import pandas as pd
import glob
import requests

# Download the stormevents databases
BRONZE_PATH = '/usr/datalake/bronze/stormevents'
SILVER_PATH = '/usr/datalake/silver/stormevents'
NCEI_URL = 'https://www.ncei.noaa.gov/pub/data/swdi/stormevents/csvfiles'

CSV_FILENAME = f'{SILVER_PATH}/docs/csvfiles.csv'

os.makedirs(f'{BRONZE_PATH}/csvfiles', exist_ok=True)

index = pd.read_csv(CSV_FILENAME)
index.head()

for _, csvfile in index.iterrows():
    filename = csvfile['filename']
    if csvfile['year'] < 1990:
        print(f'Skipping {filename}. Year earlier than 1990.')
        continue
    
    filename_local = f'{BRONZE_PATH}/csvfiles/{filename}'
    if os.path.exists(filename_local):
        print(f'Skipping {filename}. File already exists.')
        continue

    pattern = f'{BRONZE_PATH}/csvfiles/{filename[:36]}*.csv.gz'
    if len(glob.glob(pattern)):
        print(f'WARNING: Found older version of {filename}. Delete the older version.')
        continue

    url = f'{NCEI_URL}/{filename}'
    response = requests.get(url)
    response.raise_for_status()

    with open(filename_local, 'wb') as f:
        f.write(response.content)

    print(f"Downloaded {filename}.")
