==> experiments/storm_events/04_transform_data.ipynb <==
import os
import pandas as pd
import glob

BRONZE_PATH = '/usr/datalake/bronze/stormevents/csvfiles'
SILVER_PATH = '/usr/datalake/silver/stormevents/csvfiles'
PATTERN = 'StormEvents_details-ftp_v1.0_d*.csv.gz'
DTYPES = {
    'SOURCE': str, # Column 26
    'MAGNITUDE_TYPE': str, # Column 28
    'FLOOD_CAUSE': str, # Column 29
    'TOR_OTHER_WFO': str, # Column 34
    'TOR_OTHER_CZ_STATE': str, # Column 35
    'TOR_OTHER_CZ_NAME': str, # Column 37
    'EPISODE_NARRATIVE': str # Column 48
}

os.makedirs(SILVER_PATH, exist_ok=True)

search = f'{BRONZE_PATH}/{PATTERN}'

for filename in glob.glob(search):
    silver = f'{SILVER_PATH}/{filename.replace("\\","/").split("/")[-1]}'[:-3]
    if os.path.exists(silver):
        print(f"Skipping {silver} because it already exists")
        continue

    df = pd.read_csv(filename, dtype=DTYPES)
    df.to_csv(silver, index=False)
    print(f"Transformed {silver}")
