==> experiments/storm_events/12_transform_tornado_maidenhead.ipynb <==
import pandas as pd
import maidenhead as mh
import numpy as np

# Define some constants
SILVER_IN_PATH = '/usr/datalake/silver/stormevents/csvfiles/severe_dillon.csv'
SILVER_OUT_PATH = '/usr/datalake/silver/stormevents/csvfiles/tornado_chasable_date_maidenhead.csv'
use_columns = ['EFFECTIVE', 'BEGIN_LAT', 'BEGIN_LON', 'EF_0', 'EF_1', 'EF_2', 'EF_3', 'EF_4', 'EF_5']

def effective_filter(original_df: pd.DataFrame) -> pd.DataFrame:
    result = original_df.copy()
    result['EFFECTIVE'] = pd.to_datetime(result['EFFECTIVE'])
    result = result[(result['EFFECTIVE'].dt.hour < 20) & (result['EFFECTIVE'].dt.hour > 13)]
    result['EFFECTIVE'] = result['EFFECTIVE'].dt.date

    return result

def add_tornado_minor(original_df: pd.DataFrame) -> pd.DataFrame:
    result = original_df.copy()
    result['TORNADO_MINOR'] = np.minimum(result['EF_0'] + result['EF_1'], 1)
    result = result.drop(['EF_0', 'EF_1'], axis=1)

    return result

def add_tornado_significant(original_df: pd.DataFrame) -> pd.DataFrame:
    result = original_df.copy()
    result['TORNADO_SIGNIFICANT'] = np.minimum(result['EF_2'] + result['EF_3'] + result['EF_4'] + result['EF_5'], 1)
    result = result.drop(['EF_2', 'EF_3', 'EF_4', 'EF_5'], axis=1)

    return result

def tornado_filter(original_df: pd.DataFrame) -> pd.DataFrame:
    result = original_df.copy()
    result = result[(result['TORNADO_MINOR'] > 0) | (result['TORNADO_SIGNIFICANT'] > 0)]

    return result

def add_maidenhead(original_df: pd.DataFrame) -> pd.DataFrame:
    result = original_df.copy()
    result['MAIDENHEAD'] = result.apply(lambda x: mh.to_maiden(x.BEGIN_LAT, x.BEGIN_LON, 2), axis=1)
    result = result.drop(labels=['BEGIN_LAT', 'BEGIN_LON'], axis=1)

    return result

def groupby_date_maidenhead(original_df: pd.DataFrame) -> pd.DataFrame:
    result = original_df.copy()
    result = result.groupby(['EFFECTIVE', 'MAIDENHEAD']).max()
    result.loc[(result['TORNADO_MINOR'] == 1) & (result['TORNADO_SIGNIFICANT'] == 1), 'TORNADO_MINOR'] = 0

    return result

df = pd.read_csv(SILVER_IN_PATH, usecols=use_columns)
df = effective_filter(df)
df = add_tornado_minor(df)
df = add_tornado_significant(df)
df = tornado_filter(df)
df = add_maidenhead(df)
df = groupby_date_maidenhead(df)
df.head()

df.shape

df.to_csv(SILVER_OUT_PATH)
