==> experiments/storm_events/22_igra_autoencoder_index.ipynb <==
import glob
import pandas as pd

from pathlib import Path

ARTIFACTS_PATH = '/usr/datalake/silver/stormevents/artifacts/igra_storm_event_autoencoder'
STATION_LIST = '/usr/datalake/silver/igra/doc/igra2-station-list.csv'

def enumerate_brains(base_path: str) -> list[str]:
    result = []

    for path in glob.glob(f'{base_path}/*_autoencoder.pt'):
        filename = Path(path).name
        station_id = filename.removesuffix('_autoencoder.pt')

        result.append(station_id)
    
    return result

df_all_stations = pd.read_csv(STATION_LIST)
df_all_stations.head()

brains = enumerate_brains(ARTIFACTS_PATH)
brains

df_filtered_stations = df_all_stations[df_all_stations['id'].isin(brains)]
df_filtered_stations.head()

df_filtered_stations.to_csv(f'{ARTIFACTS_PATH}/brains_station_list.csv', index=False)
