==> experiments/storm_events/25_igra_encode.ipynb <==
import os
import pickle
import pandas as pd
import torch

from torch.nn import Module, Sequential, Linear, Tanh

ARTIFACTS_PATH = '/usr/datalake/silver/stormevents/artifacts/igra_storm_event_autoencoder'
STATION_LIST = '/usr/datalake/silver/igra/doc/igra2-station-list.csv'
IGRA_SOURCE = '/usr/datalake/silver/igra/gph20s10k'
IGRA_DESTINATION = '/usr/datalake/silver/stormevents/igra_encoded'

def load_data(station_id: str) -> tuple[pd.DataFrame, pd.DataFrame]:
    # Load the dataset
    filename = f'{IGRA_SOURCE}/{station_id}-data-gph20s10k.csv'
    result = pd.read_csv(filename)

    # Remove irrelevant data
    result = result[result['hour'] == 12]
    result = result.drop(['id', 'hour', '0_gph',
                '1_gph', '2_gph', '3_gph', '4_gph', '5_gph',
                '6_gph', '7_gph', '8_gph', '9_gph', '10_gph',
                '11_gph', '12_gph', '13_gph', '14_gph', '15_gph',
                '16_gph', '17_gph', '18_gph', '19_gph', '20_gph'
                ], axis=1)
    result = result.reset_index(drop=True)

    index = result.iloc[:,:2]
    data = result.iloc[:,2:]

    return index, data

def load_scaler(station_id: str):
    with open(f'{ARTIFACTS_PATH}/{station_id}_scaler.pkl', 'rb') as f:
        scaler = pickle.load(f)

    return scaler

def scale_data(scaler, sample: pd.DataFrame):
    result = torch.from_numpy(scaler.transform(sample)).float()
    return result

def load_model(station_id: str):
    model = AutoEncoder()
    model.load_state_dict(torch.load(f'{ARTIFACTS_PATH}/{station_id}_autoencoder.pt'))
    model.eval()

    return model

class AutoEncoder(Module):
    def __init__(self):
        super().__init__()
        
        self.encoder = Sequential(
            Linear(105, 80),
            Tanh(),
            Linear(80, 60),
            Tanh(),
            Linear(60, 40),
            Tanh(),
            Linear(40, 20)
        )

        self.decoder = Sequential(
            Linear(20, 40),
            Tanh(),
            Linear(40, 60),
            Tanh(),
            Linear(60, 80),
            Tanh(),
            Linear(80, 105)
        )

    def forward(self, x):
        encoded = self.encoder(x)
        decoded = self.decoder(encoded)

        return decoded

os.makedirs(IGRA_DESTINATION, exist_ok=True)

df_brains = pd.read_csv(f'{ARTIFACTS_PATH}/brains_station_list.csv')
df_brains.head()

for _, brain in df_brains.iterrows():
    station_id = brain['id']
    print(station_id)

    scaler = load_scaler(station_id)
    model = load_model(station_id)

    df_index, df_data = load_data(station_id)
    data = scale_data(scaler, df_data)
    
    # Encode the values
    logits = model.encoder(data).clone().detach().numpy()
    
    # Add the index back to the dataframe
    output = pd.DataFrame(logits)
    output = pd.concat([df_index, output], axis=1)

    output.to_csv(f'{IGRA_DESTINATION}/{station_id}_igra_encoded.csv', index=False)

output.head()
